refactor: extract ChangesPanel into focused components
